==> Part3-FlemishCap/03_tif_to_csv.ipynb <==
from osgeo import gdal
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
from raster2xyz.raster2xyz import Raster2xyz

input_raster = "hma_Present_MeanPresProbRaster_RF.tif"
out_csv = "hma_Present_MeanPresProbRaster_RF.csv"
rtxyz = Raster2xyz()
rtxyz.translate(input_raster, out_csv)
myRasterDF = pd.read_csv(out_csv)
myRasterDF

input_raster = "lma_Present_MeanPresProbRaster_RF.tif"
out_csv = "lma_Present_MeanPresProbRaster_RF.csv"
rtxyz = Raster2xyz()
rtxyz.translate(input_raster, out_csv)
myRasterDF = pd.read_csv(out_csv)
myRasterDF
